Convert object access function to take 64-bit values
Support programmatic access to values larger than 32 bits.

This most straight-forward solution; simply bumping the access
function's value pointer to *uint64_t, unfortunately breaks the API.

+++ include/co_obj.h
@@ -109,7 +109,7 @@ CO_EXPORT uint32_t co_od1001_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Pre-defined error field object (1003h)
@@ -129,7 +129,7 @@ CO_EXPORT uint32_t co_od1003_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for COB-ID SYNC message object (1005h)
@@ -149,7 +149,7 @@ CO_EXPORT uint32_t co_od1005_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Commmunication cycle object (1006h)
@@ -169,7 +169,7 @@ CO_EXPORT uint32_t co_od1006_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Synchronous window length object (1007h)
@@ -189,7 +189,7 @@ CO_EXPORT uint32_t co_od1007_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Guard time object (100Ch)
@@ -209,7 +209,7 @@ CO_EXPORT uint32_t co_od100C_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Life time factor object (100Dh)
@@ -229,7 +229,7 @@ CO_EXPORT uint32_t co_od100D_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Store Parameters object (1010h)
@@ -249,7 +249,7 @@ CO_EXPORT uint32_t co_od1010_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Restore Default Parameters object (1011h)
@@ -269,7 +269,7 @@ CO_EXPORT uint32_t co_od1011_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for COB-ID EMCY object (1014h)
@@ -289,7 +289,7 @@ CO_EXPORT uint32_t co_od1014_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Inhibit time EMCY object (1015h)
@@ -309,7 +309,7 @@ CO_EXPORT uint32_t co_od1015_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Consumer heartbeat time object (1016h)
@@ -329,7 +329,7 @@ CO_EXPORT uint32_t co_od1016_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Producer heartbeat time object (1017h)
@@ -349,7 +349,7 @@ CO_EXPORT uint32_t co_od1017_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Synchronous counter overflow value object (1019h)
@@ -369,7 +369,7 @@ CO_EXPORT uint32_t co_od1019_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Verify Configuration object (1020h)
@@ -389,7 +389,7 @@ CO_EXPORT uint32_t co_od1020_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Emergency consumer object (1028h)
@@ -409,7 +409,7 @@ CO_EXPORT uint32_t co_od1028_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for Error behavior object (1029h)
@@ -429,7 +429,7 @@ CO_EXPORT uint32_t co_od1029_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for RPDO communication parameter object (1400h - 15FFh)
@@ -449,7 +449,7 @@ CO_EXPORT uint32_t co_od1400_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for RPDO mapping parameter object (1600h - 17FFh)
@@ -469,7 +469,7 @@ CO_EXPORT uint32_t co_od1600_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for TPDO communication parameter object (1800h - 19FFh)
@@ -489,7 +489,7 @@ CO_EXPORT uint32_t co_od1800_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 /**
  * Access function for RPDO mapping parameter object (1A00h - 1BFFh)
@@ -509,7 +509,7 @@ CO_EXPORT uint32_t co_od1A00_fn (
    const co_obj_t * obj,
    const co_entry_t * entry,
    uint8_t subindex,
-   uint32_t * value);
+   uint64_t * value);
 
 #ifdef __cplusplus
 }
